feat(annotation-sync): surface Maintenance as a top-level Settings entry
Previously the Maintenance section sat at the bottom of the WebDAV settings
screen — invisible from the global Settings list. This change makes
Maintenance a sibling row under the existing "Annotations Sync" category
in Settings, drilling into a dedicated screen.

- New AnnotationSyncMaintenanceScreen + ViewModel (extracted verbatim from
  the combined screen). Devices list, Forget / Compact / Rename actions,
  confirm dialogs, and the result banner all live here.
- AnnotationSyncSettingsScreen is now purely Connection (WebDAV URL /
  user / password / test / save / clear). Its title shortens to
  "Annotation Sync · WebDAV" so the drill-down breadcrumb stays clear.
- Settings adds a second ListItem ("Maintenance") below the existing
  "WebDAV" row. Subtitle nudges to set up WebDAV first when nothing is
  configured, otherwise lists the available actions.
- New nav route ANNOTATION_SYNC_MAINTENANCE wires the row into MainScreen.

The Maintenance screen degrades gracefully when reached without a config
(explainer copy) or when no logged-in ABS server has an absUserId yet
(prompts the user to sign in).

Also picks up README cleanup: annotations are now described as
WebDAV-syncable and searchable across the library, replacing the older
"local only" line.
